Media analyzer checkpoint
