+++ core/src/main/java/io/github/chrisribble/ffm/ffmpeg/core/DecoderContext.java
@@ -1,0 +1,8 @@
+package io.github.chrisribble.ffm.ffmpeg.core;
+
+import java.lang.foreign.MemorySegment;
+
+record DecoderContext(
+		StreamInfo videoStream,
+		// AVCodecContext*
+		MemorySegment pAvCodecContext) {}
