+++ core/src/main/java/io/github/chrisribble/ffm/ffmpeg/core/FrameReceiveResult.java
@@ -1,0 +1,8 @@
+package io.github.chrisribble.ffm.ffmpeg.core;
+
+enum FrameReceiveResult {
+	READ,
+	SKIP,
+	AGAIN,
+	EOF
+}
