+++ core/src/main/java/io/github/chrisribble/ffm/ffmpeg/core/Macros.java
@@ -1,0 +1,19 @@
+package io.github.chrisribble.ffm.ffmpeg.core;
+
+/**
+ * Utility class for FFmpeg function macros
+ *
+ * This only exists because jextract does not support generating Java
+ * classes for function-style macros.
+ */
+final class Macros {
+	private Macros() {}
+
+	/**
+	 * Negate error code enum to match library return code scheme
+	 * See libavutil/error.h
+	 */
+	public static int AVERROR(final int code) {
+		return -1 * code;
+	}
+}
