+++ core/src/main/java/io/github/chrisribble/ffm/ffmpeg/core/exception/AVAllocateException.java
@@ -1,0 +1,9 @@
+package io.github.chrisribble.ffm.ffmpeg.core.exception;
+
+public class AVAllocateException extends AVException {
+	private static final long serialVersionUID = 1L;
+
+	public AVAllocateException(final String message) {
+		super(message);
+	}
+}
