+++ core/src/main/java/io/github/chrisribble/ffm/ffmpeg/core/exception/AVException.java
@@ -1,0 +1,13 @@
+package io.github.chrisribble.ffm.ffmpeg.core.exception;
+
+public class AVException extends RuntimeException {
+	private static final long serialVersionUID = 1L;
+
+	public AVException(final String message) {
+		super(message);
+	}
+
+	public AVException(final String message, final Throwable cause) {
+		super(message, cause);
+	}
+}
