+++ core/src/main/java/io/github/chrisribble/ffm/ffmpeg/core/exception/AVIOException.java
@@ -1,0 +1,13 @@
+package io.github.chrisribble.ffm.ffmpeg.core.exception;
+
+public class AVIOException extends AVException {
+	private static final long serialVersionUID = 1L;
+
+	public AVIOException(final String message) {
+		super(message);
+	}
+
+	public AVIOException(final String message, final Throwable cause) {
+		super(message, cause);
+	}
+}
