+++ core/src/main/resources/simplelogger.properties
