+++ core/src/test/java/io/github/chrisribble/ffm/ffmpeg/core/LibavVersionTest.java
@@ -1,0 +1,24 @@
+package io.github.chrisribble.ffm.ffmpeg.core;
+
+import static org.testng.Assert.assertFalse;
+import static org.testng.Assert.assertNotNull;
+import static org.testng.Assert.assertTrue;
+
+import org.testng.annotations.Test;
+
+public class LibavVersionTest {
+	@Test
+	public void getVersionInfo() {
+		String versionInfo = LibavVersion.getVersionInfo();
+		assertNotNull(versionInfo, "versionInfo should be non-null");
+		assertFalse(versionInfo.isEmpty(), "versionInfo should be non-empty");
+		assertFalse(versionInfo.isBlank(), "versionInfo should be non-blank");
+		assertTrue(versionInfo.length() >= 3, "versionInfo should contain at least 3 characters");
+	}
+
+	@Test
+	public void testIsCompatible() {
+		assertTrue(LibavVersion.getInstance().isCompatible(),
+				"runtime FFmpeg version (" + LibavVersion.getVersionInfo() + ") should be compatible with binding version");
+	}
+}
