--- example/src/main/java/io/github/chrisribble/ffmpeg/example/DecoderContext.java
@@ -1,5 +1,0 @@
-package io.github.chrisribble.ffmpeg.example;
-
-import java.lang.foreign.MemorySegment;
-
-record DecoderContext(VideoStream videoStream, MemorySegment pAvCodecContext) {}
