--- example/src/main/java/io/github/chrisribble/ffmpeg/example/FrameReceiveResult.java
@@ -1,8 +1,0 @@
-package io.github.chrisribble.ffmpeg.example;
-
-enum FrameReceiveResult {
-	READ,
-	SKIP,
-	AGAIN,
-	EOF
-}
