--- example/src/main/java/io/github/chrisribble/ffmpeg/example/exception/AVAllocateException.java
@@ -1,9 +1,0 @@
-package io.github.chrisribble.ffmpeg.example.exception;
-
-public class AVAllocateException extends AVException {
-	private static final long serialVersionUID = 1L;
-
-	public AVAllocateException(final String message) {
-		super(message);
-	}
-}
